Extract desktop shell layout module
